# backend/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"MeProfile": {
				"description": "Profile shape returned by the authenticated \"me\" endpoint.",
				"properties": {
					"userId": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"displayName": {
						"type": "string",
						"nullable": true
					},
					"userRoles": {
						"items": {
							"properties": {
								"role": {
									"properties": {
										"name": {
											"type": "string"
										}
									},
									"required": [
										"name"
									],
									"type": "object"
								}
							},
							"required": [
								"role"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"userId",
					"email",
					"displayName",
					"userRoles"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthApiResponse_MeProfile_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/MeProfile"
					},
					"accessToken": {
						"type": "string"
					},
					"requiresOnboarding": {
						"type": "boolean"
					},
					"onboardingToken": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthApiResponse_void_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"user": {},
					"accessToken": {
						"type": "string"
					},
					"requiresOnboarding": {
						"type": "boolean"
					},
					"onboardingToken": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RequestMagicLinkBody": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthApiResponse_any_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"user": {},
					"accessToken": {
						"type": "string"
					},
					"requiresOnboarding": {
						"type": "boolean"
					},
					"onboardingToken": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyMagicLinkBody": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CompleteOnboardingBody": {
				"properties": {
					"onboardingToken": {
						"type": "string"
					},
					"displayName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"GoogleLoginBody": {
				"properties": {
					"credential": {
						"type": "string"
					}
				},
				"required": [
					"credential"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RefreshTokenBody": {
				"properties": {
					"refresh_token": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"LogoutBody": {
				"properties": {
					"refresh_token": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "stock-backend",
		"version": "1.0.0",
		"description": "Stock prediction, decision making and analytics platform.",
		"contact": {
			"name": "Hamza Ahmed"
		}
	},
	"paths": {
...
		"/api/v1/auth/me": {
			"get": {
				"operationId": "GetMyInfo",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthApiResponse_MeProfile_"
								}
							}
						}
					}
				},
				"description": "Get currently authenticated user details and assigned roles.",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/auth/magic-link": {
			"post": {
				"operationId": "RequestMagicLink",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthApiResponse_void_"
								}
							}
						}
					}
				},
				"description": "Request a magic login/signup link sent via email.",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestMagicLinkBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/verify-magic-link": {
			"post": {
				"operationId": "VerifyMagicLinkToken",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthApiResponse_any_"
								}
							}
						}
					}
				},
				"description": "Verify a magic link token to log in or proceed to onboarding.",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyMagicLinkBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/onboarding": {
			"post": {
				"operationId": "CompleteOnboarding",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthApiResponse_any_"
								}
							}
						}
					}
				},
				"description": "Complete new user onboarding or update user profile.",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CompleteOnboardingBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/google": {
			"post": {
				"operationId": "GoogleLogin",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthApiResponse_any_"
								}
							}
						}
					}
				},
				"description": "Authenticate using Google OAuth 2.0 credential.",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GoogleLoginBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/refresh-token": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"accessToken": {
											"type": "string"
										},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"accessToken",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Refresh JWT access token using the HTTP-only refresh token cookie or body token.",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RefreshTokenBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Log out user and revoke refresh session.",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LogoutBody"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}